+++ sqllogictest-engines/src/postgres/simple.rs
@@ -1,4 +1,4 @@
-use std::process::{Command, ExitStatus};
+use std::process::Command;
 use std::time::Duration;
 
 use async_trait::async_trait;
@@ -66,7 +66,7 @@ impl sqllogictest::AsyncDB for Postgres<Simple> {
         tokio::time::sleep(dur).await
     }
 
-    async fn run_command(command: Command) -> std::io::Result<ExitStatus> {
-        tokio::process::Command::from(command).status().await
+    async fn run_command(command: Command) -> std::io::Result<std::process::Output> {
+        tokio::process::Command::from(command).output().await
     }
 }